+++ apps/enhanced-journal.js
@@ -147,10 +147,7 @@ export class EnhancedJournal extends Application {
             if (this.object instanceof Promise)
                 this.object = await this.object;
 
-            if(this.object.data.flags['monks-enhanced-journal']?.type)
-                this.object.data.type = this.object.data.flags['monks-enhanced-journal']?.type;
-            if (this.object.data.type == 'journalentry')
-                this.object.data.type = 'base';
+            MonksEnhancedJournal.fixType(this.object);
 
             let contentform = $('.content > section', this.element);
 
@@ -195,7 +192,7 @@ export class EnhancedJournal extends Application {
                 this.subsheet.refresh();
 
             $('.content', this.element).attr('entity-type', this.object.data.type).attr('entity-id', this.object.id);
-            contentform.empty().attr('class', this.subsheet.options.classes.concat([`${game.system.id}`]).join(' ')).append(this.subdocument);
+            contentform.empty().attr('class', this.subsheet.options.classes.join(' ')).append(this.subdocument); //.concat([`${game.system.id}`]).join(' ')
 
             //connect the tabs to the enhanced journal so that opening the regular document won't try and change tabs on the other window.
             this._tabs = this.subsheet.options.tabs.map(t => {
@@ -1055,7 +1052,7 @@ export class EnhancedJournal extends Application {
 
     async convert(type) {
         this.object._sheet = null;
-        this.object.data.type = type;
+        MonksEnhancedJournal.fixType(this.object, type);
         await this.object.setFlag('monks-enhanced-journal', 'type', type);
         await ui.sidebar.tabs.journal.render(true)
         //MonksEnhancedJournal.updateDirectory($('#journal'));